requirements.txt
ib_insync
pandas
pyyaml
streamlit
plotly
psutil
